move COO left-multiply tests into test_coo.cc
