+++ test/test_sparse_data/test_coo.cc
@@ -1,5 +1,4 @@
-#include "test/test_sparse_data/common.hh"
-#include "../comparison.hh"
+#include "test/test_sparse_data/test_left_multiply.hh"
 #include <vector>
 
 using namespace RandBLAS::sparse_data;
@@ -183,3 +182,201 @@ TEST_F(Test_SkOp_to_COO, LASO_Dim_15by7) {
     sparse_skop_to_coo(15, 7, 0, 3, RandBLAS::MajorAxis::Long);
     sparse_skop_to_coo(15, 7, 1, 3, RandBLAS::MajorAxis::Long);
 }
+
+
+template <typename T>
+class TestLeftMultiply_COO : public TestLeftMultiply_Sparse<COOMatrix<T>> {
+    COOMatrix<T> make_test_matrix(int64_t m, int64_t n, T nonzero_prob, uint32_t key = 0) {
+        randblas_require(nonzero_prob >= 0);
+        randblas_require(nonzero_prob <= 1);
+        COOMatrix<T> A(m, n);
+        std::vector<T> actual(m * n);
+        RandBLAS::RNGState s(key);
+        iid_sparsify_random_dense<T>(m, n, Layout::ColMajor, actual.data(), 1 - nonzero_prob, s);
+        dense_to_coo<T>(Layout::ColMajor, actual.data(), 0.0, A);
+        return A;
+    }
+};
+
+class TestLeftMultiply_COO_double : public TestLeftMultiply_COO<double> {};
+
+class TestLeftMultiply_COO_single : public TestLeftMultiply_COO<float> {};
+
+
+////////////////////////////////////////////////////////////////////////
+//
+//
+//      Left-muliplication
+//
+//
+////////////////////////////////////////////////////////////////////////
+
+TEST_F(TestLeftMultiply_COO_double, tall_multiply_eye_colmajor) {
+    for (uint32_t key : {0}) {
+        multiply_eye(key, 200, 30, Layout::ColMajor, 0.01);
+        multiply_eye(key, 200, 30, Layout::ColMajor, 0.10);
+        multiply_eye(key, 200, 30, Layout::ColMajor, 0.80);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_double, tall_multiply_eye_rowmajor) {
+    for (uint32_t key : {0}) {
+        multiply_eye(key, 200, 30, Layout::RowMajor, 0.01);
+        multiply_eye(key, 200, 30, Layout::RowMajor, 0.10);
+        multiply_eye(key, 200, 30, Layout::RowMajor, 0.80);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_double, wide_multiply_eye_colmajor) {
+    for (uint32_t key : {0}) {
+        multiply_eye(key, 51, 101, Layout::ColMajor, 0.01);
+        multiply_eye(key, 51, 101, Layout::ColMajor, 0.10);
+        multiply_eye(key, 51, 101, Layout::ColMajor, 0.80);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_double, wide_multiply_eye_rowmajor) {
+    for (uint32_t key : {0}) {
+        multiply_eye(key, 51, 101, Layout::RowMajor, 0.01);
+        multiply_eye(key, 51, 101, Layout::RowMajor, 0.10);
+        multiply_eye(key, 51, 101, Layout::RowMajor, 0.80);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_double, nontrivial_scales_colmajor1) {
+    double alpha = 5.5;
+    double beta = 0.0;
+    alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.05);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.10);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.80);
+}
+
+TEST_F(TestLeftMultiply_COO_double, nontrivial_scales_colmajor2) {
+    double alpha = 5.5;
+    double beta = -1.0;
+    alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.05);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.10);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.80);
+}
+
+TEST_F(TestLeftMultiply_COO_double, nontrivial_scales_rowmajor1) {
+    double alpha = 5.5;
+    double beta = 0.0;
+    alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.05);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.10);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.80);
+}
+
+TEST_F(TestLeftMultiply_COO_double, nontrivial_scales_rowmajor2) {
+    double alpha = 5.5;
+    double beta = -1.0;
+    alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.05);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.10);
+    alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.80);
+}
+
+////////////////////////////////////////////////////////////////////////
+//
+//      transpose of self (sparse operator)
+//
+////////////////////////////////////////////////////////////////////////
+
+TEST_F(TestLeftMultiply_COO_double, transpose_self_colmajor) {
+    for (uint32_t key : {0}) {
+        transpose_self(key, 200, 30, Layout::ColMajor, 0.01);
+        transpose_self(key, 200, 30, Layout::ColMajor, 0.10);
+        transpose_self(key, 200, 30, Layout::ColMajor, 0.80);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_double, transpose_self_rowmajor) {
+    for (uint32_t key : {0}) {
+        transpose_self(key, 200, 30, Layout::RowMajor, 0.01);
+        transpose_self(key, 200, 30, Layout::RowMajor, 0.10);
+        transpose_self(key, 200, 30, Layout::RowMajor, 0.80);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_single, transpose_self) {
+    for (uint32_t key : {0}) {
+        transpose_self(key, 200, 30, Layout::ColMajor, 0.01);
+        transpose_self(key, 200, 30, Layout::ColMajor, 0.10);
+        transpose_self(key, 200, 30, Layout::ColMajor, 0.80);
+    }
+}
+
+////////////////////////////////////////////////////////////////////////
+//
+//      Submatrices of self (sparse operator)
+//
+////////////////////////////////////////////////////////////////////////
+
+TEST_F(TestLeftMultiply_COO_double, submatrix_self_colmajor) {
+    for (uint32_t key : {0}) {
+        submatrix_self(key, 3, 10, 8, 12, 3, 1, Layout::ColMajor, 0.1);
+        submatrix_self(key, 3, 10, 8, 12, 3, 1, Layout::ColMajor, 1.0);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_double, submatrix_self_rowmajor) {
+    for (uint32_t key : {0}) {
+        submatrix_self(key, 3, 10, 8, 12, 3, 1, Layout::RowMajor, 0.1);
+        submatrix_self(key, 3, 10, 8, 12, 3, 1, Layout::RowMajor, 1.0);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_single, submatrix_self) {
+    for (uint32_t key : {0}) {
+        submatrix_self(key, 3, 10, 8, 12, 3, 1, Layout::ColMajor, 0.1);
+        submatrix_self(key, 3, 10, 8, 12, 3, 1, Layout::ColMajor, 1.0);
+    }
+}
+
+////////////////////////////////////////////////////////////////////////
+//
+//     submatrix of other operand in left-multiply
+//
+////////////////////////////////////////////////////////////////////////
+
+TEST_F(TestLeftMultiply_COO_double, submatrix_other_colmajor) {
+    for (uint32_t key : {0}) {
+        submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 0.1);
+        submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 1.0);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_double, submatrix_other_rowmajor) {
+    for (uint32_t key : {0}) {
+        submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::RowMajor, 0.1);
+        submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::RowMajor, 1.0);
+    }
+}
+
+TEST_F(TestLeftMultiply_COO_single, submatrix_other) {
+    for (uint32_t key : {0}) {
+        submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 0.1);
+        submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 1.0);
+    }
+}
+
+////////////////////////////////////////////////////////////////////////
+//
+//     transpose of other
+//
+////////////////////////////////////////////////////////////////////////
+
+
+TEST_F(TestLeftMultiply_COO_double, sparse_times_trans_other_colmajor) {
+    uint32_t key = 0;
+    transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.05);
+    transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.10);
+    transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.80);
+}
+
+TEST_F(TestLeftMultiply_COO_double, sparse_times_trans_other_rowmajor) {
+    uint32_t key = 0;
+    transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.05);
+    transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.10);
+    transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.80);
+}
+
